fix: unused function and setter + amount bug set
Amount will be updated later in the function so it should be set to 0,
this is not very optimized because this ask 2 calls to the database I
will think about this.

Fixed bug where no error would be raised when exceeding max capacity
and make sure that amount is not incremented if that error is raised
by placing it before that line.

Fixed max_capacity not enough to accomodate upgrades

Fixed error in logic for checking if exceeding capacity